+++ src/plugin.js
@@ -1,45 +1,245 @@
 import videojs from 'video.js';
 
-// Default options for the plugin.
-const defaults = {};
+const Html5 = videojs.getTech('Html5');
 
 /**
- * Function to invoke when the player is ready.
+ * For the most part, these are the events that occur early in the lifecycle
+ * of a player, but there is considerable variability across browsers and
+ * devices (not to mention properties like autoplay and preload). As such, we
+ * listen to a bunch of events for source changes.
  *
- * This is a great place for your plugin to initialize itself. When this
- * function is called, the player will have its DOM and child components
- * in place.
- *
- * @function onPlayerReady
- * @param    {Player} player
- * @param    {Object} [options={}]
+ * @type {Array}
  */
-const onPlayerReady = (player, options) => {
-  player.addClass('vjs-per-source-behaviors');
+const CHANGE_DETECT_EVENTS = [
+  'abort',
+  'canplay',
+  'emptied',
+  'loadeddata',
+  'loadedmetadata',
+  'loadstart',
+  'play',
+  'playing'
+];
+
+/**
+ * Whether or not the player is in an ad state. Ideally, this function would
+ * not need to exist, but hooks provided by contrib-ads are not sufficient to
+ * cover all conditions at this time.
+ *
+ * Additionally, it'd be preferable to not have to read from the DOM and have
+ * external code simply disable/enable "sourcechanged" detection when entering
+ * or leaving an ad state.
+ *
+ * @return {Boolean}
+ */
+const isInAdPlayback = (p) => p.hasClass('vjs-ad-loading') || p.hasClass('vjs-ad-playing');
+
+/**
+ * Creates an event binder function of a given type.
+ *
+ * @param  {Boolean} isOne
+ *         Rather than delegating to the player's `one()` method, we want to
+ *         retain full control over when the listener is unbound (particularly
+ *         due to the ability for per-source behaviors to be toggled on and
+ *         off at will).
+ *
+ * @return {Function}
+ */
+const createPerSrcBinder = (isOne) => {
+  return function(first, second) {
+
+    // Do not bind new listeners when per-source behaviors are disabled.
+    if (this.perSourceBehaviors.disabled()) {
+      return;
+    }
+
+    const isTargetPlayer = arguments.length === 2;
+    const originalSrc = this.currentSrc();
+
+    // This array is the set of arguments to use for `on()` and `off()` methods
+    // of the player.
+    const args = [first];
+
+    // Make sure we bind here so that a GUID is set on the original listener
+    // and that it is bound to the proper context.
+    const originalListener = videojs.bind(
+      isTargetPlayer ? this : first,
+      arguments[arguments.length - 1]
+    );
+
+    // The wrapped listener `subargs` are the arguments passed to the original
+    // listener (i.e. the Event object and an additional data hash).
+    const wrappedListener = (...subargs) => {
+      const changed = this.currentSrc() !== originalSrc;
+
+      // Do not evaluate listeners if per-source behaviors are disabled.
+      if (this.perSourceBehaviors.disabled()) {
+        return;
+      }
+
+      if (changed || isOne) {
+        this.off(...args);
+      }
+
+      if (!changed) {
+        originalListener(...subargs);
+      }
+    };
+
+    // Make sure the wrapped listener and the original listener share a GUID,
+    // so that video.js properly removes event bindings when `off()` is passed
+    // the original listener!
+    wrappedListener.guid = originalListener.guid;
+
+    // If we are targeting a different object from the player, we need to include
+    // the second argument.
+    if (!isTargetPlayer) {
+      args.push(second);
+    }
+
+    args.push(wrappedListener);
+
+    return this.on(...args);
+  };
 };
 
 /**
- * A video.js plugin.
+ * Bind an event listener on a per-source basis.
  *
- * In the plugin function, the value of `this` is a video.js `Player`
- * instance. You cannot rely on the player being in a "ready" state here,
- * depending on how the plugin is invoked. This may or may not be important
- * to you; if not, remove the wait for "ready"!
+ * @function onPerSrc
+ * @param  {String|Array|Component|Element} first
+ *         The event type(s) or target Component or Element.
+ *
+ * @param  {Function|String|Array} second
+ *         The event listener or event type(s) (when `first` is target).
+ *
+ * @param  {Function} third
+ *         The event listener (when `second` is event type(s)).
+ *
+ * @return {Player}
+ */
+const onPerSrc = createPerSrcBinder();
+
+/**
+ * Bind an event listener on a per-source basis. This listener can only
+ * be called once.
+ *
+ * @function onePerSrc
+ * @param  {String|Array|Component|Element} first
+ *         The event type(s) or target Component or Element.
+ *
+ * @param  {Function|String|Array} second
+ *         The event listener or event type(s) (when `first` is target).
+ *
+ * @param  {Function} third
+ *         The event listener (when `second` is event type(s)).
+ *
+ * @return {Player}
+ */
+const onePerSrc = createPerSrcBinder(true);
+
+/**
+ * Applies per-source behaviors to a video.js Player object.
  *
  * @function perSourceBehaviors
- * @param    {Object} [options={}]
- *           An object of options left to the plugin author to define.
  */
-const perSourceBehaviors = function(options) {
-  this.ready(() => {
-    onPlayerReady(this, videojs.mergeOptions(defaults, options));
+const perSourceBehaviors = function() {
+  let cachedSrc;
+  let disabled = false;
+  let srcChangeTimer;
+
+  this.perSourceBehaviors = {
+
+    /**
+     * Disable per-source behaviors on this player.
+     *
+     * @return {Boolean}
+     */
+    disable() {
+      window.clearTimeout(srcChangeTimer);
+      srcChangeTimer = null;
+      disabled = true;
+      return disabled;
+    },
+
+    /**
+     * Whether per-source behaviors are disabled on this player.
+     *
+     * @return {Boolean}
+     */
+    disabled() {
+      return disabled;
+    },
+
+    /**
+     * Enable per-source behaviors on this player.
+     *
+     * @return {Boolean}
+     */
+    enable() {
+      disabled = false;
+      return disabled;
+    },
+
+    /**
+     * Whether per-source behaviors are disabled on this player.
+     *
+     * @return {Boolean}
+     */
+    enabled() {
+      return !disabled;
+    },
+
+    VERSION: '__VERSION__'
+  };
+
+  // Add the per-source event binding methods to this player.
+  this.onPerSrc = onPerSrc;
+  this.onePerSrc = onePerSrc;
+
+  this.on(CHANGE_DETECT_EVENTS, (e) => {
+
+    if (
+      this.perSourceBehaviors.disabled() ||
+      srcChangeTimer ||
+      !this.currentSrc() ||
+      isInAdPlayback(this)
+    ) {
+      return;
+    }
+
+    // Track any and all interim events from this one until the next tick
+    // when we evaluate the timer.
+    const interimEvents = [];
+
+    const addInterimEvent = (f) => {
+      interimEvents.push({time: Date.now(), event: f});
+    };
+
+    addInterimEvent(e);
+    this.on(Html5.Events, addInterimEvent);
+
+    srcChangeTimer = window.setTimeout(() => {
+      let currentSrc = this.currentSrc();
+
+      this.off(Html5.Events, addInterimEvent);
+      srcChangeTimer = null;
+
+      if (currentSrc && currentSrc !== cachedSrc) {
+
+        this.trigger('sourcechanged', {
+          interimEvents,
+          from: cachedSrc,
+          to: currentSrc
+        });
+
+        cachedSrc = currentSrc;
+      }
+    }, 1);
   });
 };
 
-// Register the plugin with video.js.
 videojs.plugin('perSourceBehaviors', perSourceBehaviors);
-
-// Include the version number.
 perSourceBehaviors.VERSION = '__VERSION__';
 
 export default perSourceBehaviors;