+++ CHANGELOG.md
@@ -2,7 +2,7 @@ CHANGELOG
 =========
 
 ## HEAD (Unreleased)
-_(none)_
+* Added `"sourcechanged"` event to a video.js player.
+* Added `onPerSrc()` and `onePerSrc()` methods to a video.js player.
 
 --------------------
-